- arrays: 1. findStringBestMatch: a. Add RegExp type to candidates type-constraint. b. Remove redundant check if a candidate is of type string. c. Delete console logging line used for testing during development. d. Fix tests to reflect changes.
- fs:
1. isValidDirEntry:
a. Refactor implementation logic to fix test coverage not covering 100%.
2. readDirSyncWithTypes:
a. Refactor if-statement to be compatible with isValidDirEntry new changes.
2. readDirSyncWithTypes:
a. Refactor if-statement to be compatible with isValidDirEntry new changes.

- propchecker:
1. isOfType:
a. Fix logic of inner function isValid to better handle type = 'null'.
b. Now returns false if type/types is not of type string, function, or object.
2. Property:
a. Refactor setter for property caller to fix test coverage.
b. Add tests to cover class Property.
3. PropertyParams:
a. Refactor type of property caller to use exported type PrimitiveTypes and add type string.
b. Refactor type of property callerClass to use exported type Constructor and add type string.
4. PropertyTypeError:
a. Add tests to cover class PropertyTypeError.
5. PropertyRequiredTypeError:
a. Add tests to cover class PropertyRequiredTypeError.

- strings:
1. machSubstrings:
a. Add test to cover an uncovered branch in the logic of the function.